fix: add types and length to decorators, reorganize file structure and maintain PascalCase for all file names across the app
