+++ src/controllers/UserController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { AppDataSource } from '../data-source';
+import { AppDataSource } from '../database/DataSource';
 import { User } from '../entities/User';
 
 export class UserController {
