+++ src/entities/User.ts
@@ -1,4 +1,4 @@
-import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, OneToOne } from 'typeorm';
+import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, OneToOne, DeleteDateColumn } from 'typeorm';
 import { Auth } from './Auth';
 
 export enum UserRole {
@@ -9,18 +9,18 @@ export enum UserRole {
 @Entity()
 export class User {
   @PrimaryGeneratedColumn()
-  id!: number;
+  userId!: number;
 
-  @Column({ unique: true })
+  @Column({ unique: true, type: 'varchar', length: 50 })
   username!: string;
 
-  @Column()
+  @Column({type: 'varchar', length: 100})
   name!: string;
 
-  @Column({ unique: true })
+  @Column({ unique: true, type: 'varchar', length: 100 })
   email!: string;
 
-  @CreateDateColumn()
+  @CreateDateColumn({ type: 'timestamp' })
   joinDate!: Date;
 
   @Column({
@@ -35,4 +35,7 @@ export class User {
 
   @OneToOne(() => Auth, (auth: Auth) => auth.user)
   auth!: Auth;
+
+  @DeleteDateColumn({ type: 'timestamp', nullable: true })
+  deletedAt!: Date | null;
 }
